Some modifications in Offers parsing logic

+++ system/library/hobotix/Amazon/OffersParser.php
@@ -418,12 +418,14 @@ class OffersParser
 
 			if ($supplier = $this->Suppliers->getSupplier($rfOffer->getSellerName())){
 				if (!empty($supplier['amzn_coefficient']) && (int)$supplier['amzn_coefficient'] < $this->Suppliers->supplierMinInnerRatingForUse){
+					echo('amzn_coefficient');
 					$addThisOffer = false;
 				}
 			}
 
 			if ((float)$this->config->get('config_rainforest_max_delivery_price') && (float)$rfOffer->getDeliveryAmount() > 0){
 				if ((float)$rfOffer->getDeliveryAmount() > $this->config->get('config_rainforest_max_delivery_price')){
+					echo('config_rainforest_max_delivery_price');
 					$addThisOffer = false;
 				}
 			}
@@ -431,6 +433,7 @@ class OffersParser
 			//Bad delivery price
 			if ((float)$this->config->get('config_rainforest_max_delivery_price_multiplier') && (float)$rfOffer->getDeliveryAmount() > 0){
 				if ((float)$rfOffer->getDeliveryAmount() > (float)$rfOffer->getPriceAmount() * (float)$this->config->get('config_rainforest_max_delivery_price_multiplier')){
+					echo('config_rainforest_max_delivery_price_multiplier');
 					$addThisOffer = false;
 				}
 			}
@@ -438,12 +441,14 @@ class OffersParser
 			if ($rfOffer->getDeliveryComments() && $offerDates = $this->parseAmazonDeliveryComment($rfOffer->getDeliveryComments())){
 				if ($this->config->get('config_rainforest_max_delivery_days_for_offer') > 0){
 					if (!empty($offerDates['minDays']) && (int)$offerDates['minDays'] > (int)$this->config->get('config_rainforest_max_delivery_days_for_offer')){
+						echo('config_rainforest_max_delivery_days_for_offer');
 						$addThisOffer = false;
 					}
 				}
 			}
 
 			if ($rfOffer->getSellerRating50()>0 && $rfOffer->getSellerRating50() < $this->Suppliers->supplierMinRatingForUse){
+				echo('supplierMinRatingForUse');
 				$addThisOffer = false;
 			}
 
@@ -485,16 +490,31 @@ class OffersParser
 
 		if ((int)$this->config->get('config_rainforest_skip_low_price_products') > 0){
 			if ((int)$amazonBestPrice > 0 && (int)$amazonBestPrice < (int)$this->config->get('config_rainforest_skip_low_price_products')){
-				//do the action
+
 
 				foreach ($products as $product_id){
-					if ($this->PriceLogic->checkIfProductIsInOrders($product_id)){
-						$this->model_product_edit->disableProduct($product_id)->addAsinToIgnored($asin, $this->model_product_get->getProductName($product_id));
-						return 'disabled';
-					} else {
-						$this->model_product_edit->deleteProductSimple($product_id)->addAsinToIgnored($asin, $this->model_product_get->getProductName($product_id));
-						return 'deleted';
+					$is_added_from_amazon = $this->model_product_get->getProductsByAsin($product_id);
+
+					if ($is_added_from_amazon && $this->config->get('config_rainforest_drop_low_price_products')){
+						if ($this->PriceLogic->checkIfProductIsInOrders($product_id)){
+							$this->model_product_edit->disableProduct($product_id)->addAsinToIgnored($asin, $this->model_product_get->getProductName($product_id));
+							return 'disabled';
+						} else {
+							$this->model_product_edit->deleteProductSimple($product_id)->addAsinToIgnored($asin, $this->model_product_get->getProductName($product_id));
+							return 'deleted';
+						}
 					}
+
+					if (!$is_added_from_amazon && $this->config->get('config_rainforest_drop_low_price_products_for_manual')){
+						if ($this->PriceLogic->checkIfProductIsInOrders($product_id)){
+							$this->model_product_edit->disableProduct($product_id)->addAsinToIgnored($asin, $this->model_product_get->getProductName($product_id));
+							return 'disabled';
+						} else {
+							$this->model_product_edit->deleteProductSimple($product_id)->addAsinToIgnored($asin, $this->model_product_get->getProductName($product_id));
+							return 'deleted';
+						}
+					}
+
 				}
 
 			}
